[BREAKING CHANGES] Backend endpoint routes refactor
