+++ backend/config/mongodb/data/admins.js
@@ -1,0 +1,10 @@
+module.exports = [
+  {
+    _id: '614ce88d8101e980a49f0429',
+    emailAddress: 'admin@example.com',
+    password: 'Admin123',
+    firstName: 'Admin',
+    lastName: 'User',
+    registrationDate: '2023-10-23T00:00:00Z',
+  },
+];
