+++ backend/config/mongodb/data/users.js
@@ -3,7 +3,7 @@ module.exports = [
     _id: '614ce88d8101e980a49f0427',
     emailAddress: 'user1@example.com',
     phoneNumber: 1234567890,
-    password: 'Password123',
+    password: 'Asdasd42',
     firstName: 'John',
     lastName: 'Doe',
     registrationDate: '2023-10-23T00:00:00Z',
@@ -17,7 +17,7 @@ module.exports = [
     _id: '614ce88d8101e980a49f0428',
     emailAddress: 'user2@example.com',
     phoneNumber: 9876543210,
-    password: 'Securepass123',
+    password: 'Asdasd42',
     firstName: 'Alice',
     lastName: 'Smith',
     registrationDate: '2023-10-23T00:00:00Z',
@@ -29,13 +29,13 @@ module.exports = [
   },
   {
     _id: '614ce88d8101e980a49f0429',
-    emailAddress: 'admin@example.com',
+    emailAddress: 'lalo@example.com',
     phoneNumber: 5555555555,
-    password: 'Admin123',
-    firstName: 'Admin',
+    password: 'Asdasd42',
+    firstName: 'Lalo',
     lastName: 'User',
     registrationDate: '2023-10-23T00:00:00Z',
-    role: 'admin',
+    role: 'user',
     twoFactorEnabled: true,
     active: true,
     emailConfirmed: true,
@@ -45,7 +45,7 @@ module.exports = [
     _id: '614ce88d8101e980a49f0430',
     emailAddress: 'jane@example.com',
     phoneNumber: 9998887777,
-    password: 'Janepassword123',
+    password: 'Asdasd42',
     firstName: 'Jane',
     lastName: 'Johnson',
     registrationDate: '2023-10-23T00:00:00Z',
@@ -59,7 +59,7 @@ module.exports = [
     _id: '614ce88d8101e980a49f0431',
     emailAddress: 'inactive@example.com',
     phoneNumber: 1112223333,
-    password: 'Inactivepass123',
+    password: 'Asdasd42',
     firstName: 'Inactive',
     lastName: 'User',
     registrationDate: '2023-10-23T00:00:00Z',
