+++ backend/routes/admin/authRoutes.js
@@ -1,0 +1,20 @@
+const express = require('express');
+const authController = require('../../controllers/admin/authController');
+const authValidator = require('../../validators/admin/authValidator');
+
+const router = express.Router();
+
+router.post('/signin', authValidator.signin, authController.signin);
+router.post(
+  '/password/forgot',
+  authValidator.forgotPassword,
+  authController.forgotPassword,
+);
+router.patch(
+  '/password/reset',
+  authValidator.resetPassword,
+  authController.resetPassword,
+);
+router.post('/verify-2fa', authController.verify2fa);
+
+module.exports = router;
